# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlexScale",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api",
      "variables": {}
    }
  ],
  "paths": {
    "/auth/verifyToken":{
      "get": {
        "tags": [
          "Login"
        ],
        "summary": "verifyToken",
        "operationId": "verifyToken",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "TOKEN"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/organization": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "getAllOrganizations",
        "operationId": "getAllOrganizations",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "summary": "createOrganization",
        "operationId": "createOrganization",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/createOrganizationRequest"
                  },
                  {
                    "example": {
                      "Name": "Amazon",
                      "CreatedBy": "6386106b8534e766fcbada14",
                      "BillingContactEmail": "amazon@test.com",
                      "BillingContactPhone": 854612387,
                      "BillingContactFirstName": "NopCommerce",
                      "BillingContactLastName": "Template",
                      "BillingAddress": "test address",
                      "BillingCity": "Karachi",
                      "BillingState": "Sindh",
                      "BillingCountry": "Pakistan",
                      "BillingPostalCode": "8541",
                      "OrganizationType": "Client",
                      "BillingAddress2": "test BillingAddress 2",
                      "Address1": "test address 1",
                      "Address2": "test address 2",
                      "City": "Karachi",
                      "State": "Sindh",
                      "PostalCode": "8541",
                      "Country": "Pakistan"
                    }
                  }
                ]
              },
              "example": {
                "Name": "Amazon",
                "CreatedBy": "6386106b8534e766fcbada14",
                "BillingContactEmail": "amazon@test.com",
                "BillingContactPhone": 854612387,
                "BillingContactFirstName": "NopCommerce",
                "BillingContactLastName": "Template",
                "BillingAddress": "test address",
                "BillingCity": "Karachi",
                "BillingState": "Sindh",
                "BillingCountry": "Pakistan",
                "BillingPostalCode": "8541",
                "OrganizationType": "Client",
                "BillingAddress2": "test BillingAddress 2",
                "Address1": "test address 1",
                "Address2": "test address 2",
                "City": "Karachi",
                "State": "Sindh",
                "PostalCode": "8541",
                "Country": "Pakistan"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Organization"
        ],
        "summary": "updateOrganization",
        "operationId": "updateOrganization",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63b3ce620a243766b6cb4ed8"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/updateOrganizationRequest"
                  },
                  {
                    "example": {
                      "Name": "Amazon",
                      "BillingContactEmail": "amazon@test.com",
                      "BillingContactPhone": 854612387,
                      "BillingContactFirstName": "NopCommerce",
                      "BillingContactLastName": "Template",
                      "BillingAddress": "test BillingAddress",
                      "BillingCity": "Karachi",
                      "BillingState": "Sindh",
                      "BillingPostalCode": "8541",
                      "BillingCountry": "Canada",
                      "OrganizationType": "Client",
                      "BillingAddress2": "test BillingAddress 2",
                      "Address1": "test address 1",
                      "Address2": "test address 2",
                      "City": "Karachi",
                      "State": "Sindh",
                      "PostalCode": "8541",
                      "Country": "Pakistan"
                    }
                  }
                ]
              },
              "example": {
                "Name": "Amazon",
                "BillingContactEmail": "amazon@test.com",
                "BillingContactPhone": 854612387,
                "BillingContactFirstName": "NopCommerce",
                "BillingContactLastName": "Template",
                "BillingAddress": "test BillingAddress",
                "BillingCity": "Karachi",
                "BillingState": "Sindh",
                "BillingPostalCode": "8541",
                "BillingCountry": "Canada",
                "OrganizationType": "Client",
                "BillingAddress2": "test BillingAddress 2",
                "Address1": "test address 1",
                "Address2": "test address 2",
                "City": "Karachi",
                "State": "Sindh",
                "PostalCode": "8541",
                "Country": "Pakistan"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      } ,
      "delete": {
        "tags": [
          "Organization"
        ],
        "summary": "deleteOrganization",
        "operationId": "deleteOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63b3ce620a243766b6cb4ed8"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      }      
    },
...
    "/document": {
      "post": {
        "tags": [
          "Document"
        ],
        "summary": "uploadDoc",
        "operationId": "uploadDoc",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/uploadDocRequest"
                  },
                  {
                    "example": {
                      "Location": "https://flexscalebucket.s3-us-west-2.amazonaws.com/uploads/63b6ad27e6aad9c1d9c621b8/Test.png",
                      "User": "63b6ad27e6aad9c1d9c621b8",
                      "prevDocIds": [
                        "63bbc09cf7fa506de853b795"
                      ],
                      "FileName": "Test",
                      "CreatedBy": "6386106b8534e766fcbada14"
                    }
                  }
                ]
              },
              "example": {
                "Location": "https://flexscalebucket.s3-us-west-2.amazonaws.com/uploads/63b6ad27e6aad9c1d9c621b8/Test.png",
                "User": "63b6ad27e6aad9c1d9c621b8",
                "prevDocIds": [
                  "63bbc09cf7fa506de853b795"
                ],
                "FileName": "Test",
                "CreatedBy": "6386106b8534e766fcbada14"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "put": {
        "tags": [
          "Document"
        ],
        "summary": "editDoc",
        "operationId": "editDoc",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63bbcbf1f7fa506de853c357"
            }
          },
          {
            "name": "FileName",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "Test Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      },
      "delete": {
        "tags": [
          "Document"
        ],
        "summary": "deleteDoc",
        "operationId": "deleteDoc",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "example": "63bbcbf1f7fa506de853c357"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [{
            "bearerAuth": []
        }]
      }
    }
  },
  "components": {
    "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
    },
    "schemas": {
...
      "createOrganizationRequest": {
        "title": "createOrganizationRequest",
        "required": [
          "Name",
          "CreatedBy",
          "BillingContactEmail",
          "BillingContactPhone",
          "BillingContactFirstName",
          "BillingContactLastName",
          "BillingAddress",
          "BillingCity",
          "BillingState",
          "BillingCountry",
          "BillingPostalCode",
          "OrganizationType",
          "BillingAddress2",
          "Address1",
          "Address2",
          "City",
          "State",
          "PostalCode",
          "Country"
        ],
        "type": "object",
        "properties": {
          "Name": {
            "type": "string"
          },
          "CreatedBy": {
            "type": "string"
          },
          "BillingContactEmail": {
            "type": "string"
          },
          "BillingContactPhone": {
            "type": "integer",
            "format": "int32"
          },
          "BillingContactFirstName": {
            "type": "string"
          },
          "BillingContactLastName": {
            "type": "string"
          },
          "BillingAddress": {
            "type": "string"
          },
          "BillingCity": {
            "type": "string"
          },
          "BillingState": {
            "type": "string"
          },
          "BillingCountry": {
            "type": "string"
          },
          "BillingPostalCode": {
            "type": "string"
          },
          "OrganizationType": {
            "type": "string"
          },
          "BillingAddress2": {
            "type": "string"
          },
          "Address1": {
            "type": "string"
          },
          "Address2": {
            "type": "string"
          },
          "City": {
            "type": "string"
          },
          "State": {
            "type": "string"
          },
          "PostalCode": {
            "type": "string"
          },
          "Country": {
            "type": "string"
          }
        },
        "example": {
          "Name": "Amazon",
          "CreatedBy": "6386106b8534e766fcbada14",
          "BillingContactEmail": "amazon@test.com",
          "BillingContactPhone": 854612387,
          "BillingContactFirstName": "NopCommerce",
          "BillingContactLastName": "Template",
          "BillingAddress": "test address",
          "BillingCity": "Karachi",
          "BillingState": "Sindh",
          "BillingCountry": "Pakistan",
          "BillingPostalCode": "8541",
          "OrganizationType": "Client",
          "BillingAddress2": "test BillingAddress 2",
          "Address1": "test address 1",
          "Address2": "test address 2",
          "City": "Karachi",
          "State": "Sindh",
          "PostalCode": "8541",
          "Country": "Pakistan"
        }
      },
      "updateOrganizationRequest": {
        "title": "updateOrganizationRequest",
        "required": [
          "Name",
          "BillingContactEmail",
          "BillingContactPhone",
          "BillingContactFirstName",
          "BillingContactLastName",
          "BillingAddress",
          "BillingCity",
          "BillingState",
          "BillingPostalCode",
          "BillingCountry",
          "OrganizationType",
          "BillingAddress2",
          "Address1",
          "Address2",
          "City",
          "State",
          "PostalCode",
          "Country"
        ],
        "type": "object",
        "properties": {
          "Name": {
            "type": "string"
          },
          "BillingContactEmail": {
            "type": "string"
          },
          "BillingContactPhone": {
            "type": "integer",
            "format": "int32"
          },
          "BillingContactFirstName": {
            "type": "string"
          },
          "BillingContactLastName": {
            "type": "string"
          },
          "BillingAddress": {
            "type": "string"
          },
          "BillingCity": {
            "type": "string"
          },
          "BillingState": {
            "type": "string"
          },
          "BillingPostalCode": {
            "type": "string"
          },
          "BillingCountry": {
            "type": "string"
          },
          "OrganizationType": {
            "type": "string"
          },
          "BillingAddress2": {
            "type": "string"
          },
          "Address1": {
            "type": "string"
          },
          "Address2": {
            "type": "string"
          },
          "City": {
            "type": "string"
          },
          "State": {
            "type": "string"
          },
          "PostalCode": {
            "type": "string"
          },
          "Country": {
            "type": "string"
          }
        },
        "example": {
          "Name": "Amazon",
          "BillingContactEmail": "amazon@test.com",
          "BillingContactPhone": 854612387,
          "BillingContactFirstName": "NopCommerce",
          "BillingContactLastName": "Template",
          "BillingAddress": "test BillingAddress",
          "BillingCity": "Karachi",
          "BillingState": "Sindh",
          "BillingPostalCode": "8541",
          "BillingCountry": "Canada",
          "OrganizationType": "Client",
          "BillingAddress2": "test BillingAddress 2",
          "Address1": "test address 1",
          "Address2": "test address 2",
          "City": "Karachi",
          "State": "Sindh",
          "PostalCode": "8541",
          "Country": "Pakistan"
        }
      },
...
      "uploadDocRequest": {
        "title": "uploadDocRequest",
        "required": [
          "Location",
          "User",
          "prevDocIds",
          "FileName",
          "CreatedBy"
        ],
        "type": "object",
        "properties": {
          "Location": {
            "type": "string"
          },
          "User": {
            "type": "string"
          },
          "prevDocIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": ""
          },
          "FileName": {
            "type": "string"
          },
          "CreatedBy": {
            "type": "string"
          }
        },
        "example": {
          "Location": "https://flexscalebucket.s3-us-west-2.amazonaws.com/uploads/63b6ad27e6aad9c1d9c621b8/Test.png",
          "User": "63b6ad27e6aad9c1d9c621b8",
          "prevDocIds": [
            "63bbc09cf7fa506de853b795"
          ],
          "FileName": "Test",
          "CreatedBy": "6386106b8534e766fcbada14"
        }
      }
    }
  },
...
}